CMRARC-544 removing old files, updating tests, moving e-mail organizer to lib folder.
