--- app/mailers/curator_mailer.rb
@@ -1,13 +1,0 @@
-class CuratorMailer < ApplicationMailer
-
-  def released_records(users, records)
-#    @records = records
-#    mail(to: users.map(&:email), subject: 'Metadata Curation Tool Released Records')
-  end
-
-  def closed_records(users, records)
-#    @records = records
-#    mail(to: users.map(&:email), subject: 'Metadata Curation Tool Closed Records')
-  end
-
-end
