--- app/views/curator_mailer/closed_records.html.erb
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-
-  <head>
-    <meta content='text/html; charset=UTF-8' http-equiv='Content-Type' />
-    <%= render partial: 'shared/google_tag_manager_head' %>
-  </head>
-
-  <body>
-  <%= render partial: 'shared/google_tag_manager_body' %>
-
-    <h1>Metadata Curation Records Are Closed</h1>
-    <p>The following Metadata Curation records have been closed:</p>
-
-    <ul>
-      <% @records.each do |r| %>
-        <li><%= r.concept_id %></li>
-      <% end %>
-    </ul>
-
-    <p>Please sign in to the Metadata Curation Tool to view these records.</p>
-  </body>
-
-</html>
