--- app/views/curator_mailer/closed_records.text.erb
@@ -1,9 +1,0 @@
-Metadata Curation Records Are Closed
-
-The following Metadata Curation records have been closed:
-
-<% @records.each do |r| %>
-  <%= r.concept_id %>
-<% end %>
-
-Please sign in to the Metadata Curation Tool to view these records.
