--- app/views/curator_mailer/released_records.html.erb
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-
-  <head>
-    <meta content='text/html; charset=UTF-8' http-equiv='Content-Type' />
-    <%= render partial: 'shared/google_tag_manager_head' %>
-  </head>
-
-  <body>
-  <%= render partial: 'shared/google_tag_manager_body' %>
-
-    <h1>Records Are Waiting on Your Review</h1>
-    <p>The following Metadata Curation records have been made available for your review:</p>
-
-    <ul>
-      <% @records.each do |r| %>
-        <li><%= r.concept_id %></li>
-      <% end %>
-    </ul>
-
-    <p>Please sign in to the Metadata Curation Tool to review these records.</p>
-  </body>
-
-</html>
