--- app/views/curator_mailer/released_records.text.erb
@@ -1,9 +1,0 @@
-Metadata Review Records are Waiting on Your Review
-
-The following Metadata Curation records have been made available for your review:
-
-<% @records.each do |r| %>
-  <%= r.concept_id %>
-<% end %>
-
-Please sign in to the Metadata Curation Tool to review these records.
