+++ test/mailers/daac_curator_mailer_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require "#{Rails.root}/lib/rake_helpers/daac_curator_mailer_helper"
 
 class DaacCuratorMailerTest < ActionMailer::TestCase
 
@@ -15,6 +16,68 @@ class DaacCuratorMailerTest < ActionMailer::TestCase
       assert_equal 'Summary of Reports Available to OB_DAAC', email.subject
       assert_equal read_fixture('released_records_digest_notification_text_part').join, email.text_part.body.to_s
       assert_equal read_fixture('released_records_digest_notification_html_part').join, email.html_part.body.to_s
+    end
+  end
+  
+  describe 'finding all the users and records to send digest e-mails' do
+    describe 'when two different daacs need e-mails' do
+      before do
+        @user1 = User.find(8)
+        @user1.daac = 'NSIDC'
+        @user1.save
+        @email_count = ActionMailer::Base.deliveries.count
+      end
+
+      after do
+        @user1.daac = 'OB_DAAC'
+        @user1.save
+      end
+
+      it 'sends two e-mails when two DAACs have records in review' do
+        # This time is a second Monday, so it should trigger weekly/daily emails
+        DaacCuratorMailerOrchestration.released_records_digest_conductor(now: Time.new(2020, 2, 10))
+
+        assert_equal @email_count + 2, ActionMailer::Base.deliveries.count
+        # The e-mails are not consistently sent in the same order
+        # Either, the first e-mail should match user 7 and the second match 8 or
+        # the first should match 8 and the second 7
+        assert (ActionMailer::Base.deliveries[@email_count].to == [User.find(7).email] && ActionMailer::Base.deliveries[@email_count + 1].to == [User.find(8).email]) ||
+               (ActionMailer::Base.deliveries[@email_count].to == [User.find(8).email] && ActionMailer::Base.deliveries[@email_count + 1].to == [User.find(7).email])
+      end
+    end
+
+    describe 'when users with different preferences are processed' do
+      before do
+        @email_count = ActionMailer::Base.deliveries.count
+      end
+
+      it 'sends one email to one recipient on a Tuesday' do
+        DaacCuratorMailerOrchestration.released_records_digest_conductor(now: Time.new(2020, 2, 4))
+
+        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
+        assert_equal [User.find(7).email], ActionMailer::Base.deliveries[@email_count].to
+      end
+
+      it 'sends one email to four recipients on the first Monday' do
+        DaacCuratorMailerOrchestration.released_records_digest_conductor(now: Time.new(2020, 2, 3))
+
+        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
+        assert_equal [User.find(7).email, User.find(8).email, User.find(9).email, User.find(10).email], ActionMailer::Base.deliveries[@email_count].to
+      end
+
+      it 'sends one email to two recipients on the second Monday' do
+        DaacCuratorMailerOrchestration.released_records_digest_conductor(now: Time.new(2020, 2, 10))
+
+        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
+        assert_equal [User.find(7).email, User.find(8).email], ActionMailer::Base.deliveries[@email_count].to
+      end
+
+      it 'sends one email to three recipients on the third Monday' do
+        DaacCuratorMailerOrchestration.released_records_digest_conductor(now: Time.new(2020, 2, 17))
+
+        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
+        assert_equal [User.find(7).email, User.find(8).email, User.find(9).email], ActionMailer::Base.deliveries[@email_count].to
+      end
     end
   end
 end
