+++ test/models/user_test.rb
@@ -84,63 +84,4 @@ class UserTest < ActiveSupport::TestCase
       assert user.email_preference == nil
     end
   end
-
-  describe 'finding all the users and records to send digest e-mails' do
-    describe 'when two different daacs need e-mails' do
-      before do
-        @user1 = User.find(8)
-        @user1.daac = 'NSIDC'
-        @user1.save
-        @email_count = ActionMailer::Base.deliveries.count
-      end
-
-      after do
-        @user1.daac = 'OB_DAAC'
-        @user1.save
-      end
-
-      it 'sends two e-mails when two DAACs have records in review' do
-        # This time is a second Monday, so it should trigger weekly/daily emails
-        User.released_records_digest_conductor(now: Time.new(2020, 2, 10))
-
-        assert_equal @email_count + 2, ActionMailer::Base.deliveries.count
-        assert_equal [User.find(7).email], ActionMailer::Base.deliveries[@email_count].to
-        assert_equal [User.find(8).email], ActionMailer::Base.deliveries[@email_count + 1].to
-      end
-    end
-
-    describe 'when users with different preferences are processed' do
-      before do
-        @email_count = ActionMailer::Base.deliveries.count
-      end
-
-      it 'sends one email to one recipient on a Tuesday' do
-        User.released_records_digest_conductor(now: Time.new(2020, 2, 4))
-
-        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
-        assert_equal [User.find(7).email], ActionMailer::Base.deliveries[@email_count].to
-      end
-
-      it 'sends one email to four recipients on the first Monday' do
-        User.released_records_digest_conductor(now: Time.new(2020, 2, 3))
-
-        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
-        assert_equal [User.find(7).email, User.find(8).email, User.find(9).email, User.find(10).email], ActionMailer::Base.deliveries[@email_count].to
-      end
-
-      it 'sends one email to two recipients on the second Monday' do
-        User.released_records_digest_conductor(now: Time.new(2020, 2, 10))
-
-        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
-        assert_equal [User.find(7).email, User.find(8).email], ActionMailer::Base.deliveries[@email_count].to
-      end
-
-      it 'sends one email to three recipients on the third Monday' do
-        User.released_records_digest_conductor(now: Time.new(2020, 2, 17))
-
-        assert_equal @email_count + 1, ActionMailer::Base.deliveries.count
-        assert_equal [User.find(7).email, User.find(8).email, User.find(9).email], ActionMailer::Base.deliveries[@email_count].to
-      end
-    end
-  end
 end
